requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
diploma-frontend @ file:///app/diploma-frontend-0.6.tar.gz
Django==5.2.4
djangorestframework==3.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
requests==2.32.5
